+++ android/src/main/java/com/pichillilorenzo/flutter_inappwebview/chrome_custom_tabs/ChromeCustomTabsSettings.java
@@ -9,6 +9,7 @@ import androidx.browser.trusted.ScreenOrientation;
 import androidx.browser.trusted.TrustedWebActivityDisplayMode;
 
 import com.pichillilorenzo.flutter_inappwebview.ISettings;
+import com.pichillilorenzo.flutter_inappwebview.types.AndroidResource;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -25,6 +26,12 @@ public class ChromeCustomTabsSettings implements ISettings<ChromeCustomTabsActiv
   public Boolean showTitle = true;
   @Nullable
   public String toolbarBackgroundColor;
+  @Nullable
+  public String navigationBarColor;
+  @Nullable
+  public String navigationBarDividerColor;
+  @Nullable
+  public String secondaryToolbarColor;
   public Boolean enableUrlBarHiding = false;
   public Boolean instantAppsEnabled = false;
   public String packageName;
@@ -35,6 +42,8 @@ public class ChromeCustomTabsSettings implements ISettings<ChromeCustomTabsActiv
   public List<String> additionalTrustedOrigins = new ArrayList<>();
   public TrustedWebActivityDisplayMode displayMode = null;
   public Integer screenOrientation = ScreenOrientation.DEFAULT;
+  public List<AndroidResource> startAnimations = new ArrayList<>();
+  public List<AndroidResource> exitAnimations = new ArrayList<>();
 
   @NonNull
   @Override
@@ -58,6 +67,15 @@ public class ChromeCustomTabsSettings implements ISettings<ChromeCustomTabsActiv
           break;
         case "toolbarBackgroundColor":
           toolbarBackgroundColor = (String) value;
+          break;
+        case "navigationBarColor":
+          navigationBarColor = (String) value;
+          break;
+        case "navigationBarDividerColor":
+          navigationBarDividerColor = (String) value;
+          break;
+        case "secondaryToolbarColor":
+          secondaryToolbarColor = (String) value;
           break;
         case "enableUrlBarHiding":
           enableUrlBarHiding = (Boolean) value;
@@ -102,6 +120,24 @@ public class ChromeCustomTabsSettings implements ISettings<ChromeCustomTabsActiv
         case "screenOrientation":
           screenOrientation = (Integer) value;
           break;
+        case "startAnimations":
+          List<Map<String, Object>> startAnimationsList = (List<Map<String, Object>>) value;
+          for (Map<String, Object> startAnimation : startAnimationsList) {
+            AndroidResource androidResource = AndroidResource.fromMap(startAnimation);
+            if (androidResource != null) {
+              startAnimations.add(AndroidResource.fromMap(startAnimation));
+            }
+          }
+          break;
+        case "exitAnimations":
+          List<Map<String, Object>> exitAnimationsList = (List<Map<String, Object>>) value;
+          for (Map<String, Object> exitAnimation : exitAnimationsList) {
+            AndroidResource androidResource = AndroidResource.fromMap(exitAnimation);
+            if (androidResource != null) {
+              exitAnimations.add(AndroidResource.fromMap(exitAnimation));
+            }
+          }
+          break;
       }
     }
 
@@ -115,6 +151,9 @@ public class ChromeCustomTabsSettings implements ISettings<ChromeCustomTabsActiv
     options.put("addDefaultShareMenuItem", addDefaultShareMenuItem);
     options.put("showTitle", showTitle);
     options.put("toolbarBackgroundColor", toolbarBackgroundColor);
+    options.put("navigationBarColor", navigationBarColor);
+    options.put("navigationBarDividerColor", navigationBarDividerColor);
+    options.put("secondaryToolbarColor", secondaryToolbarColor);
     options.put("enableUrlBarHiding", enableUrlBarHiding);
     options.put("instantAppsEnabled", instantAppsEnabled);
     options.put("packageName", packageName);