+++ example/integration_test/chrome_safari_browser/custom_action_button.dart
@@ -13,10 +13,12 @@ void customActionButton() {
           TargetPlatform.android,
         ].contains(defaultTargetPlatform);
 
-  test('add custom action button', () async {
+  test('add custom action button and update icon', () async {
     var chromeSafariBrowser = MyChromeSafariBrowser();
     var actionButtonIcon =
         await rootBundle.load('test_assets/images/flutter-logo.png');
+    var actionButtonIcon2 =
+        await rootBundle.load('test_assets/images/flutter-logo.jpg');
     chromeSafariBrowser.setActionButton(ChromeSafariBrowserActionButton(
         id: 1,
         description: 'Action Button description',
@@ -25,15 +27,18 @@ void customActionButton() {
     expect(chromeSafariBrowser.isOpened(), false);
 
     await chromeSafariBrowser.open(url: TEST_URL_1);
-    await chromeSafariBrowser.browserCreated.future;
+    await chromeSafariBrowser.opened.future;
     expect(chromeSafariBrowser.isOpened(), true);
     expect(() async {
       await chromeSafariBrowser.open(url: TEST_CROSS_PLATFORM_URL_1);
     }, throwsA(isInstanceOf<ChromeSafariBrowserAlreadyOpenedException>()));
 
     await expectLater(chromeSafariBrowser.firstPageLoaded.future, completes);
+    await chromeSafariBrowser.updateActionButton(
+        icon: actionButtonIcon2.buffer.asUint8List(),
+        description: 'New Action Button description');
     await chromeSafariBrowser.close();
-    await chromeSafariBrowser.browserClosed.future;
+    await chromeSafariBrowser.closed.future;
     expect(chromeSafariBrowser.isOpened(), false);
   }, skip: shouldSkip);
 }