+++ example/integration_test/chrome_safari_browser/custom_menu_item.dart
@@ -18,7 +18,7 @@ void customMenuItem() {
     expect(chromeSafariBrowser.isOpened(), false);
 
     await chromeSafariBrowser.open(url: TEST_URL_1);
-    await chromeSafariBrowser.browserCreated.future;
+    await chromeSafariBrowser.opened.future;
     expect(chromeSafariBrowser.isOpened(), true);
     expect(() async {
       await chromeSafariBrowser.open(url: TEST_CROSS_PLATFORM_URL_1);
@@ -26,7 +26,7 @@ void customMenuItem() {
 
     await expectLater(chromeSafariBrowser.firstPageLoaded.future, completes);
     await chromeSafariBrowser.close();
-    await chromeSafariBrowser.browserClosed.future;
+    await chromeSafariBrowser.closed.future;
     expect(chromeSafariBrowser.isOpened(), false);
   }, skip: shouldSkip);
 }