+++ example/integration_test/chrome_safari_browser/main.dart
@@ -6,6 +6,7 @@ import 'custom_menu_item.dart';
 import 'custom_tabs.dart';
 import 'open_and_close.dart';
 import 'trusted_web_activity.dart';
+import 'sf_safari_view_controller.dart';
 
 void main() {
   final shouldSkip =
@@ -17,5 +18,6 @@ void main() {
     customActionButton();
     customTabs();
     trustedWebActivity();
+    sfSafariViewController();
   }, skip: shouldSkip);
 }