+++ lib/src/chrome_safari_browser/main.dart
@@ -1,4 +1,8 @@
 export 'chrome_safari_browser.dart';
-export 'chrome_safari_browser_settings.dart';
+export 'chrome_safari_browser_settings.dart'
+    show
+        ChromeSafariBrowserOptions,
+        ChromeSafariBrowserSettings,
+        ChromeSafariBrowserClassOptions;
 export 'android/main.dart';
 export 'apple/main.dart';