+++ CHANGELOG.md
@@ -3,13 +3,16 @@
 - Added `headers`, `otherLikelyURLs` arguments on `ChromeSafariBrowser.open` method for Android
 - Added `onNavigationEvent`, `onServiceConnected`, `onRelationshipValidationResult` events on `ChromeSafariBrowser` for Android
 - Added `mayLaunchUrl`, `launchUrl`, `updateActionButton`, `validateRelationship` methods on `ChromeSafariBrowser` for Android
+- Added `startAnimations`, `exitAnimations`, `navigationBarColor`, `navigationBarDividerColor`, `secondaryToolbarColor` ChromeSafariBrowser settings for Android
 - Added `didLoadSuccessfully` optional argument on `ChromeSafariBrowser.onCompletedInitialLoad` event for iOS
 - Added `onInitialLoadDidRedirect`, `onWillOpenInBrowser` events on `ChromeSafariBrowser` for iOS
 - Added `clearWebsiteData`, `prewarmConnections`, `invalidatePrewarmingToken` static methods on `ChromeSafariBrowser` for iOS
+- Added `getVariationsHeader` WebView static method
 
 ### BREAKING CHANGES
 
 - `ChromeSafariBrowser.onCompletedInitialLoad` event has an optional argument
+- All `ChromeSafariBrowserSettings` properties are optionals
 
 ## 6.0.0-beta.8
 
@@ -78,7 +81,7 @@
 - Added `PullToRefreshController.isEnabled` method
 - Updated `getMetaThemeColor` on iOS 15.0+
 - Deprecated `onLoadError` for `onReceivedError`. `onReceivedError` will be called also for subframes
-- Deprecated `onLoadHttpError` for `onReceivedError`. `onReceivedHttpError` will be called also for subframes
+- Deprecated `onLoadHttpError` for `onReceivedHttpError`. `onReceivedHttpError` will be called also for subframes
 
 ### BREAKING CHANGES